// src/app/models/influencer.model.ts
export class Influencer {
    id?: any;
    userId?: any;
    influencerId?: any;
    country?: any;
    state?: any;
    city?: any;
    instagramAvatar?: any;
    instagramFollowers?: any;
    instagramPosts?: any;
    segments?: any;
    startValue?: any;
    createdAt?: any;
    updatedAt?: any;
}